requirements.txt
Django==1.8.1
django-rq==0.8.0
rq-dashboard
django-flat-theme==1.1.1
django-annoying==0.8.0
django-extensions==1.1.1
django-filepicker==0.1.4
django-registration-redux==1.2
django-ses==0.4.1
vendor/django-tastypie-0.12.2-dev.tar.gz
django-uuidfield==0.4.0
gunicorn==18.0
mimeparse==0.1.3
psycopg2==2.5.1
python-dateutil==1.5
python-postmark==0.4.4
raven==5.8.1
bernhard
django-redis==4.2.0
hiredis==0.2.0
misaka==1.0.2
readability-lxml==0.3.0.3
cssselect==0.9.1
tinycss2==0.4
html2text==2014.4.5
buffpy==1.07
bleach

boto==2.9.9
requests==2.7.0
wsgiref==0.1.2
bpython
google-api-python-client==1.2

pillow==2.8.1

--find-links https://code.stripe.com
stripe==1.9.2

fabric

django-debug-toolbar==1.0.1
pep8
django-cprofile-middleware==0.1
psutil==2.2.0
pyquery==1.2.9

elasticsearch==1.4.0
django-pgjson==0.2.3
jsonpatch==1.11
jsonschema==2.5.1
functools32

# requests SNI
pyopenssl
ndg-httpsclient
pyasn1

pymobiledetect==1.2.2
